refactor: rename plugins/ → agent-plugins/
Makes the directory name unambiguous — "plugins" is used generically
by opencode and other tools; "agent-plugins" clearly signals these are
the vibe-station AgentPlugin implementations.

Updated all import paths in routes/sessions.ts, routes/worktrees.ts,
routes/modes.ts, and the __tests__ files.
