+++ apps/cli/src/daemon/agent-plugins/claude.ts
