+++ apps/cli/src/daemon/agent-plugins/claudeRestore.ts
