+++ apps/cli/src/daemon/agent-plugins/cursor.ts
