+++ apps/cli/src/daemon/agent-plugins/cursorRestore.ts
