+++ apps/cli/src/daemon/agent-plugins/opencode.ts
