+++ apps/cli/src/daemon/agent-plugins/registry.ts
